Add micropip

+++ pyodide_build/buildpkg.py
@@ -40,6 +40,9 @@ def check_checksum(path, pkg):
 
 
 def download_and_extract(buildpath, packagedir, pkg, args):
+    srcpath = buildpath / packagedir
+    if 'source' not in pkg:
+        return srcpath
     tarballpath = buildpath / Path(pkg['source']['url']).name
     if not tarballpath.is_file():
         try:
@@ -50,7 +53,6 @@ def download_and_extract(buildpath, packagedir, pkg, args):
         except Exception:
             tarballpath.unlink()
             raise
-    srcpath = buildpath / packagedir
     if not srcpath.is_dir():
         shutil.unpack_archive(str(tarballpath), str(buildpath))
     return srcpath
@@ -65,7 +67,7 @@ def patch(path, srcpath, pkg, args):
     pkgdir = path.parent.resolve()
     os.chdir(srcpath)
     try:
-        for patch in pkg['source'].get('patches', []):
+        for patch in pkg.get('source', {}).get('patches', []):
             subprocess.run([
                 'patch', '-p1', '--binary', '-i', pkgdir / patch
             ], check=True)
@@ -73,7 +75,7 @@ def patch(path, srcpath, pkg, args):
         os.chdir(orig_dir)
 
     # Add any extra files
-    for src, dst in pkg['source'].get('extras', []):
+    for src, dst in pkg.get('source', {}).get('extras', []):
         shutil.copyfile(pkgdir / src, srcpath / dst)
 
     with open(srcpath / '.patched', 'wb') as fd:
@@ -183,9 +185,10 @@ def build_package(path, args):
     try:
         if not needs_rebuild(pkg, path, buildpath):
             return
-        if buildpath.resolve().is_dir():
-            shutil.rmtree(buildpath)
-        os.makedirs(buildpath)
+        if 'source' in pkg:
+            if buildpath.resolve().is_dir():
+                shutil.rmtree(buildpath)
+            os.makedirs(buildpath)
         srcpath = download_and_extract(buildpath, packagedir, pkg, args)
         patch(path, srcpath, pkg, args)
         compile(path, srcpath, pkg, args)
